log-consumer/requirements.txt
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-pubsub==2.21.3
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
MarkupSafe==2.1.5
packaging==24.0
pika==1.3.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
# pywin32==306
pywin32==306; sys_platform == 'win32'
requests==2.32.3
rsa==4.9
urllib3==2.2.1
Werkzeug==3.0.3
